Add support for SSL boom'ing
This just checks the URL scheme for SSL and if that exists, simply
leaves the hostname intact so it doesn't break the certificate checking.

I've also added unit tests for the resolve function. This seems to
work from my testing. Not quite sure what the motiviation to resolve
hostnames was in the first place was, so I may be breakign something.